+++ server/impl/src/test/java/com/walmartlabs/concord/server/process/queue/ProcessQueueDaoTest.java
@@ -53,7 +53,7 @@ public class ProcessQueueDaoTest extends AbstractDaoTest {
         UUID orgId = OrganizationManager.DEFAULT_ORG_ID;
 
         String projectName = "project_" + System.currentTimeMillis();
-        UUID projectId = projectDao.insert(orgId, projectName, null, null, null, null, true);
+        UUID projectId = projectDao.insert(orgId, projectName, null, null, null, null, true, new byte[0]);
 
         UUID instanceA = UUID.randomUUID();
         queueDao.insertInitial(instanceA, ProcessKind.DEFAULT, null, projectId, "testInitiator");