+++ server/impl/src/test/java/com/walmartlabs/concord/server/template/ProjectDaoTest.java
@@ -57,7 +57,7 @@ public class ProjectDaoTest extends AbstractDaoTest {
 
         Map<String, Object> cfg = ImmutableMap.of("a", "a-v");
         String projectName = "project#" + System.currentTimeMillis();
-        UUID projectId = projectDao.insert(orgId, projectName, "test", null, cfg, null, true);
+        UUID projectId = projectDao.insert(orgId, projectName, "test", null, cfg, null, true, new byte[0]);
 
         // ---
         Map<String, Object> actualCfg = projectDao.getConfiguration(projectId);
@@ -107,8 +107,8 @@ public class ProjectDaoTest extends AbstractDaoTest {
         String aName = "aProject#" + System.currentTimeMillis();
         String bName = "bProject#" + System.currentTimeMillis();
 
-        projectDao.insert(orgId, aName, "test", null, null, null, true);
-        projectDao.insert(orgId, bName, "test", null, null, null, true);
+        projectDao.insert(orgId, aName, "test", null, null, null, true, new byte[0]);
+        projectDao.insert(orgId, bName, "test", null, null, null, true, new byte[0]);
 
         // ---